# rodriguez-carlos-DelilahResto-1.0.0-swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/rodriguez-carlos/DelilahResto/1.0.0
info:
  description: This is a simple API
  version: "1.0.0"
  title: Delilah Restó API
  contact:
    email: carlosdanielrod@gmail.com
paths:
...
  /orders:
    post:
      operationId: postOrder
      summary: Creates a new order with a default status_order value of 'new'. User most be logged in to operate. orderId will be automatically generated
      requestBody:
        description: A JSON object containing order information. 
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method_order: 
                  type: string
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
              required:
              - payment_method_order
              - products
      responses:
        '200':
          description: Returns the order with its data
        '400':
          description: Forbidden
    get:
      operationId: getAllOrders
      summary: Retrieves all the orders that are authorized for the logged user. Admins see all orders, non admins only see their own orders. The info is displayed with products, user name, user address, and more.
      security: 
      - BearerAuth: []
      responses:
        '200':
          description: Returns all authorized orders
        '400':
          description: User is not an admin
          
    patch:
      operationId: patchOrderStatus
      summary: Only admins can use this operation to update the status of an order. The status_order and the id_order must be provided in a JSON object.
      security: 
      - BearerAuth: [admin]
      requestBody:
        description: A JSON object containing order information. 
        content:
          application/json:
            schema:
              type: object
              properties: 
                status_order:
                  type: string
                id_order:
                  type: integer
      responses:
        '200':
          description: Returns the new status of the order
        '400': 
          description: User is not an admin
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...